Fix to avoid the failure in CIT
This commit includes changes to 'feature-cttest' and 'module-cttest' bash commands.
Using the VSI_IMAGE_NAME for the sanbdox_uuid caused failure in 'ops-gate' jobs.
make devenv_ct_test indirectly uses the UUID and looks for the 5th field.
This was missing because of echoing the VSI_IAMGE_NAME to sandbox uuid in special cases
i.e 'ops-gate-808' had only three fields in the name and devenv_ct_test was expecting 5 fields
separated by '-'
This change will use the VSI_IMAGE_NAME for exporting the docker image only and prevent
over-writing the sandbox_uuid

Change-Id: I982d54eb22ed63e4e405682f4c0b459b5f698383
Signed-off-by: Vasanth Viswanathan <vasanth.viswanathan@hpe.com>