Remove OPS build and check jobs for topo-common
Tags: chg, dev

This repo does not need to run the OPS image build check or gate
jobs as the code in this repo is not part of the OPS image.
Running the OPS check and gate jobs forced the need to have a
recipe for the repo as the template for check/gate (in macros.yaml)
needs to "add" and "build" the repo using the make devenv_ targets.

The declaration of the topology-common jobs (.yaml) has been left
in place as a place holder as job specifics to test the code are
going to be added soon.

Change-Id: Iec385a2514a1c0c78b31ffa34f6960402c9d3f1b
Signed-off-by: Walinton Cambronero <walinton.cambronero@hpe.com>