Configuration file for ovs repo
The repo is openswitch/ovs. It's different from other repos in the sense
that it mirrors https://github.com/openvswitch/ovs, so the branches
(master, branch-2.5, etc.) are mirrors of the corresponding branches in
GitHub. The branches in the patches namespace are meant to keep the
series of patches to be applied against each branch from upstream.
Maintainers must be able to for pushes to these branches in order to be
able to reorder commits, or in general rebase the patches against
specific upstream commits.

Registered Users should be able to open code reviews for new patches on
top of the existing series and ovs-maintainers can submit reviewed code
to branches.

There's no CIT for this repo, all the CIT validations happen in
ops-build once the patches are generated and committed there.

Change-Id: I9ce75f8509d29e073df2ea0a5d61f1f70bad62f3
Signed-off-by: Marcelo E. Magallon <marcelo.magallon@hpe.com>