Use internal OVS repo with patches for apidocgen
ops-restd gate process generates the Swagger documentation using
apidocgen. This scripts uses the ops-openvswitch python libraries to
validate the content of the schema. However, ops-openvswitch needs some
patches applied by Yocto in build time (the code it was using was not
patched and therefore apidocgen.py was failing).

This fix make the swagger task use the openvswitch/ovs repo, which has
the patches applied.

Change-Id: Ie8ead53002aa482ce66c6c9494a71b896a499491
Signed-off-by: Javier Albornoz <javier.albornoz@hpe.com>